requirements.txt
Django==1.9
django-widget-tweaks==1.4.2
mysqlclient==1.3.12
Pillow==5.0.0
python-decouple==3.1
django-registration-redux==2.4
django-widget-tweaks==1.4.2
